backend/ai/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
scikit-learn
pandas
numpy
statsmodels
joblib
